+++ src/features/network/components/SearchBar.tsx
@@ -10,7 +10,7 @@ import { useForegroundColor } from '@/design-system/color/useForegroundColor';
 import { opacity } from '@/framework/ui/utils/opacity';
 import * as i18n from '@/languages';
 
-import { networkSwitcherStore } from '../stores/networkSwitcher';
+import { networkSwitcherStore } from '../stores/networkSwitcherStore';
 
 export const SEARCH_BAR_HEIGHT = 48;
 const SEARCH_BAR_BORDER_RADIUS = SEARCH_BAR_HEIGHT / 2;
