+++ src/features/network/stores/networkSwitcherStore.ts
